+++ utils/app/plugins.ts
@@ -1,9 +1,10 @@
 import {PluginKey} from "@/types/plugin"
+import {localStorageSafeGetItem, localStorageSafeRemoveItem, localStorageSafeSetItem} from "@/utils/app/storage"
 
 export const STORAGE_KEY_PLUGIN_KEYS = "pluginKeys"
 
 export const getPluginKeys = (): PluginKey[] => {
-  const PluginKeyAsString = localStorage.getItem(STORAGE_KEY_PLUGIN_KEYS)
+  const PluginKeyAsString = localStorageSafeGetItem(STORAGE_KEY_PLUGIN_KEYS)
   try {
     return PluginKeyAsString ? (JSON.parse(PluginKeyAsString) as PluginKey[]) : []
   } catch (error) {
@@ -13,8 +14,8 @@ export const getPluginKeys = (): PluginKey[] => {
 }
 
 export const savePluginKeys = (pluginKeys: PluginKey[]) =>
-  localStorage.setItem(STORAGE_KEY_PLUGIN_KEYS, JSON.stringify(pluginKeys))
+  localStorageSafeSetItem(STORAGE_KEY_PLUGIN_KEYS, JSON.stringify(pluginKeys))
 
 export const removePluginKeys = () => {
-  localStorage.removeItem(STORAGE_KEY_PLUGIN_KEYS)
+  localStorageSafeRemoveItem(STORAGE_KEY_PLUGIN_KEYS)
 }